requirements.txt
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dominate==2.9.1
Flask==3.0.3
Flask-Bootstrap4==4.0.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pycparser==2.22
requests==2.32.3
urllib3==2.2.2
visitor==0.1.3
Werkzeug==3.0.3
gunicorn
